backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
chromadb==0.5.5
sentence-transformers==3.1.1
google-generativeai==0.8.1
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
